+++ src/app/modules/v1/models/CheckParamsTrait.php
@@ -7,6 +7,11 @@ use yii\web\BadRequestHttpException;
 
 trait CheckParamsTrait
 {
+    public function exceptParams()
+    {
+        return ['page', 'fields', 'expand'];
+    }
+
     /**
      * @throws BadRequestHttpException
      */
@@ -19,7 +24,7 @@ trait CheckParamsTrait
             }
         }
         foreach ($this->params as $key => $param) {
-            if(!in_array($key, $rulesFields)) {
+            if(!in_array($key, $rulesFields) && !in_array($key, $this->exceptParams())) {
                 throw new BadRequestHttpException('Wrong params');
             }
         }